ios compile issue

+++ mobile/ios/NSControl/NSControlApp.swift
@@ -274,9 +274,9 @@ struct WebViewContainer: UIViewRepresentable {
                 if let remap = dict["remap"] as? [[String: Any]], remap.count == 3 {
                     let a0 = remap[0]; let a1 = remap[1]; let a2 = remap[2]
                     ns_set_motion_remap(
-                        intValue(a0["axis"], 1), intValue(a0["sign"], 1),
-                        intValue(a1["axis"], 2), intValue(a1["sign"], -1),
-                        intValue(a2["axis"], 0), intValue(a2["sign"], 1))
+                        Int32(intValue(a0["axis"], 1)), Int32(intValue(a0["sign"], 1)),
+                        Int32(intValue(a1["axis"], 2)), Int32(intValue(a1["sign"], -1)),
+                        Int32(intValue(a2["axis"], 0)), Int32(intValue(a2["sign"], 1)))
                 }
             case "openTouch":
                 BridgeManager.shared.disconnect()
